refactor(app): extract functions and consts
