+++ src/constants.ts
@@ -1,1 +1,4 @@
+import maplibregl from "maplibre-gl";
+
 export const MAPTILER_KEY = "Lr1q5wml4SB6TlSXbQrJ";
+export const SCENE_ORIGIN = new maplibregl.LngLat(2.370572296723708, 48.884197153165246);
