+++ src/style.css
@@ -56,19 +56,15 @@ h1 {
   z-index: 2;
   top: 20px;
   left: 20px;
-  width: 100%;
-  max-width: 400px;
   padding: 10px;
   border-radius: 3px;
   font-size: 12px;
   text-align: left;
   color: #222;
-  background: #ffffff3c;
-  pointer-events: none;
+  background: #ffffff6e;
+  border-radius: 10px;
 }
-#toolbar-content {
-  pointer-events: all; /* Allow text to be selected */
-}
+
 #toolbar-content *::selection {
   background-color: #ff0000;
   color: #fff;
